pyproject.toml
[project]
name = "ai-commerce-platform"
version = "0.1.0"
description = "AI Commerce Platform backend built with FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.30",
    "psycopg[binary]>=3.1.19",
    "alembic>=1.13.1",
    "pydantic-settings>=2.3.1",
    "redis>=5.0.4",
    "python-json-logger>=2.0.7",
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.0",
    "email-validator>=2.1.0",
    "boto3>=1.34.0",
    "rq>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.7",
    "ruff>=0.4.4",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
